expose UserOptions
