+++ src/cpp/builder/base.h
@@ -3,6 +3,7 @@
 #pragma once
 
 #include <arrow/api.h>
+#include <pgeon.h>
 
 #include <map>
 #include <memory>
@@ -44,17 +45,6 @@ struct SqlTypeInfo {
 
   // for StructBuilder
   FieldVector field_builders;
-};
-
-struct UserOptions {
-  bool string_as_dictionaries = false;
-  int default_numeric_precision = 22;  // TODO(xav) max precision of 128 decimal ?
-  int default_numeric_scale = 6;
-  int monetary_fractional_precision = 2;  // TODO(xav) lc_monetary
-
-  struct UserOptions static Defaults() {
-    return UserOptions();
-  }
 };
 
 }  // namespace pgeon
