+++ src/cpp/pg_interface.h
@@ -10,9 +10,11 @@
 
 namespace pgeon {
 
-std::shared_ptr<ArrayBuilder> MakeColumnBuilder(PGconn* conn, Oid oid, int mod);
+std::shared_ptr<ArrayBuilder> MakeColumnBuilder(PGconn* conn, Oid oid, int mod,
+                                                const UserOptions& options);
 
-std::shared_ptr<TableBuilder> MakeTableBuilder(PGconn* conn, const char* query);
+std::shared_ptr<TableBuilder> MakeTableBuilder(PGconn* conn, const char* query,
+                                               const UserOptions& options);
 
 void CopyQuery(PGconn* conn, const char* query, std::shared_ptr<TableBuilder> builder);
